refactor: united board and column components
