+++ src/components/controller/board.controller.ts
@@ -8,6 +8,14 @@ class BoardController {
     async getColumns() {
         await boardModel.getColumns();
     }
+
+    async deleteColumnById(id: string) {
+        await boardModel.deleteColumnById(id);
+    }
+
+    async updateColumnById(columnId: string, title: string) {
+        await boardModel.updateColumnById(columnId, title);
+    }
 }
 
 export default new BoardController();
