+++ src/components/view/board.view.ts
@@ -1,8 +1,7 @@
+import { Column } from '../../spa/types';
 import { state } from '../../store/state';
 import { Control } from '../../utils/Control';
 import boardController from '../controller/board.controller';
-// eslint-disable-next-line import/no-cycle
-import columnView from './column.view';
 
 class BoardView {
     async render() {
@@ -29,7 +28,7 @@ class BoardView {
         await boardController.getColumns();
         if (state.columns) {
             state.columns.forEach((elem) => {
-                columns.element.append(columnView.render(elem));
+                columns.element.append(this.renderColumn(elem));
             });
         }
         createColumn.append(columns.element);
@@ -56,6 +55,47 @@ class BoardView {
         });
 
         return board.element;
+    }
+
+    renderColumn(column: Column) {
+        const columnBox = new Control<HTMLElement>('div', 'column');
+        const columnTitle = new Control<HTMLElement>('div', 'column__title');
+        const columnName = new Control<HTMLInputElement>('input', 'column__title_name');
+        const columnRemove = new Control<HTMLElement>('div', 'column__title_remove');
+        const columnRemoveImg = new Control<HTMLImageElement>('img', 'column__title_remove-img');
+        const tasks = new Control<HTMLElement>('div', 'column__tasks');
+        const addTask = new Control<HTMLElement>('a', 'column__add-task');
+
+        columnTitle.append(columnBox.element);
+        columnName.append(columnTitle.element);
+        columnRemove.append(columnTitle.element);
+        columnRemoveImg.append(columnRemove.element);
+        tasks.append(columnBox.element);
+        addTask.append(columnBox.element);
+
+        columnName.element.value = column.title;
+        columnRemoveImg.element.src = '../assets/icons/remove-task.png';
+        addTask.element.innerHTML = 'Add task..';
+
+        columnName.element.addEventListener('focus', () => {
+            columnName.element.select();
+            columnName.element.setSelectionRange(0, 99999);
+            // eslint-disable-next-line @typescript-eslint/no-misused-promises
+            document.addEventListener('keyup', async (event) => {
+                if (event.code === 'Enter') columnName.element.blur();
+            });
+            // eslint-disable-next-line @typescript-eslint/no-misused-promises
+            columnName.element.addEventListener('focusout', async () => {
+                await boardController.updateColumnById(column._id, columnName.element.value);
+            });
+        });
+
+        // eslint-disable-next-line @typescript-eslint/no-misused-promises
+        columnRemove.element.addEventListener('click', async () => {
+            await boardController.deleteColumnById(column._id);
+            await this.update();
+        });
+        return columnBox.element;
     }
 
     async update() {
